Refactor: modulize/slim components, endless defs
