--- app/components/business_component.rb
@@ -1,17 +1,0 @@
-# frozen_string_literal: true
-
-class BusinessComponent < ViewComponent::Base
-  attr_reader :business
-
-  def initialize(business:, **options)
-    super
-    @business = business
-    @options = options
-  end
-
-  private
-
-  def saved
-    @options.fetch(:saved, false)
-  end
-end
