--- app/components/carousel_component.rb
@@ -1,4 +1,0 @@
-# frozen_string_literal: true
-
-class CarouselComponent < ViewComponent::Base
-end
