+++ app/components/core/carousel_component.html.erb
