+++ app/components/core/carousel_component.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+module Core
+  class CarouselComponent < ViewComponent::Base
+  end
+end
