+++ app/components/core/dashboard_component.html.erb
@@ -2,9 +2,9 @@
   <%= content_tag :div, class: 'flex text-sm' do %>
     <%= link_to 'SKEDADDLE', root_path, data: { turbo_frame: 'home' }, class: 'm-2 font-bold' %>
     <%= link_to 'About', about_path, data: { turbo_frame: 'home' }, class: 'm-2' %>
-    <%= link_to 'Itineraries', itineraries_path, data: { turbo_frame: 'home' }, class: 'm-2' if logged_in %>
-    <%= button_to 'Sign Out', destroy_user_session_path, method: :delete, class: 'm-2' if logged_in %>
-    <%= unless logged_in
+    <%= link_to 'Itineraries', itineraries_path, data: { turbo_frame: 'home' }, class: 'm-2' if @logged_in %>
+    <%= button_to 'Sign Out', destroy_user_session_path, method: :delete, class: 'm-2' if @logged_in %>
+    <%= unless @logged_in
           button_to 'Sign In', user_google_oauth2_omniauth_authorize_path, class: 'm-2', data: { turbo: false }
         end %>
   <% end %>
