--- app/components/dashboard_component.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-class DashboardComponent < ViewComponent::Base
-  attr_reader :logged_in
-
-  def initialize(logged_in: false)
-    super
-    @logged_in = logged_in
-  end
-end
