--- app/components/park_component.html.erb
@@ -1,9 +1,0 @@
-<%= content_tag :div, id: 'park', class: 'p-2' do %>
-  <%= image_tag park.thumbnail, class: 'image-format m-2' if park.thumbnail.present? %>
-  <%= link_to park.name, park.url, class: 'font-bold m-2' %>
-  <p class="m-2"><b>Location: </b><%= park.location %></p>
-  <p class="m-2"><b>Description: </b><%= truncate(park.description, length: 1000) %></p>
-  <p class="m-2"><b>Directions: </b><%= park.directions %></p>
-  <p class="m-2"><b>Activities: </b><%= park.activities %></p>
-  <%= button_to "Remove #{park.name}", park_path(park), method: :delete, class: 'm-2' if saved %>
-<% end %>
