--- app/components/park_component.rb
@@ -1,17 +1,0 @@
-# frozen_string_literal: true
-
-class ParkComponent < ViewComponent::Base
-  attr_reader :park
-
-  def initialize(park:, **options)
-    super
-    @park = park
-    @options = options
-  end
-
-  private
-
-  def saved
-    @options.fetch(:saved, false)
-  end
-end
