--- app/components/place_component.html.erb
@@ -1,4 +1,0 @@
-<%= content_tag :div, id: place.main, class: 'w-1/2 md:w-1/4 p-4' do %>
-  <p class='font-bold'><%= place.name %></p>
-  <p><%= place.address %></p>
-<% end %>
