+++ app/components/sub/business_component.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Sub
+  class BusinessComponent < ViewComponent::Base
+    def initialize(business:, saved: false)
+      super
+      @business = business
+      @saved = saved
+    end
+  end
+end
