+++ app/components/sub/category_component.html.erb
@@ -1,0 +1,6 @@
+<% if @items&.any? %>
+  <%= content_tag :div, id: @category.downcase do %>
+    <h3><b><%= @category.capitalize %></b></h3>
+    <%= render Sub::BusinessComponent.with_collection(@items, saved: @saved) %>
+  <% end %>
+<% end %>
