+++ app/components/sub/category_component.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+module Sub
+  class CategoryComponent < ViewComponent::Base
+    def initialize(items:, category:, saved: false)
+      super
+      @items = items
+      @category = category
+      @saved = saved
+    end
+  end
+end
