+++ app/components/sub/park_component.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Sub
+  class ParkComponent < ViewComponent::Base
+    def initialize(park:, saved: false)
+      super
+      @park = park
+      @saved = saved
+    end
+  end
+end
