+++ app/views/home/_welcome.html.erb
@@ -16,6 +16,6 @@
     <% end %>
   <% end %>
   <%= content_tag :div, class: 'hidden lg:block lg:w-2/5' do %>
-    <%= render CarouselComponent.new %>
+    <%= render Core::CarouselComponent.new %>
   <% end %>
 <% end %>
