+++ app/views/home/about.html.erb
@@ -9,13 +9,14 @@
           <br>
           <p>This app is the continued solo exploration of a group project for the Backend Engineering program at Turing School of Software & Design.</p>
           <br>
-          <p>Got any questions? Please don't hesitate to <%= link_to 'contact', contact_path,
-                                                                     class: 'font-bold text-slate-600 hover:text-slate-700' %> us directly.</p>
+          <p>Got any questions? Please don't hesitate to
+            <%= link_to 'contact', contact_path, class: 'font-bold text-slate-600 hover:text-slate-700' %>
+            us directly.</p>
         </div>
       <% end %>
     <% end %>
   <% end %>
   <%= content_tag :div, class: 'hidden lg:block lg:w-2/5' do %>
-    <%= render CarouselComponent.new %>
+    <%= render Core::CarouselComponent.new %>
   <% end %>
 <% end %>
