+++ app/views/itineraries/index.html.erb
@@ -25,6 +25,6 @@
     <% end %>
   <% end %>
   <%= content_tag :div, class: 'hidden lg:block lg:w-2/5' do %>
-    <%= render CarouselComponent.new %>
+    <%= render Core::CarouselComponent.new %>
   <% end %>
 <% end %>
