+++ app/views/itineraries/new.html.erb
@@ -1,3 +1,3 @@
-<%= render ItineraryComponent.new search: params[:search],
-                                  city: @geocode[:city],
-                                  items: @items %>
+<%= render Core::ItineraryComponent.new search: params[:search],
+                                        city: @geocode[:city],
+                                        items: @items %>
