+++ app/views/layouts/application.html.erb
@@ -23,7 +23,7 @@
   <body>
     <div id="background-image"></div>
 
-    <%= render DashboardComponent.new logged_in: current_user %>
+    <%= render Core::DashboardComponent.new logged_in: current_user %>
 
     <div class="container mx-auto p-10 text-slate-700">
       <%= yield %>
