+++ spec/factories/businesses.rb
@@ -4,10 +4,9 @@ FactoryBot.define do
   factory :business do
     name { Faker::Business.name }
     rating { rand(5) }
-    price { '$' }
-    thumbnail { Faker::Lorem.word }
     url { Faker::Lorem.word }
-    categories { Faker::Lorem.words(number: 2) }
-    phone { Faker::PhoneNumber.cell_phone }
+    categories { Faker::Lorem.word }
+    location { Faker::Address.city }
+    url { Faker::Internet.url }
   end
 end
