+++ spec/factories/parks.rb
@@ -3,13 +3,10 @@
 FactoryBot.define do
   factory :park do
     name { Faker::Address.community }
-    city { Faker::Address.city }
-    state { Faker::Address.state }
-    country { Faker::Address.country }
+    location { Faker::Address.city }
     description { Faker::Lorem.sentence }
     directions { Faker::Lorem.sentence }
-    lat { Faker::Address.latitude }
-    lon { Faker::Address.longitude }
     activities { Faker::Lorem.words(number: 2) }
+    url { Faker::Internet.url }
   end
 end
